feat(chat): add support for updating state with position and radius
This commit adds support for updating the state of the chat with the user's position and radius. It introduces a new message type `ClientUpdateStateMessage` and a corresponding protocol function `toClientUpdateStateMessage` to create the message. The `useChat` hook has been updated to send this message to the server when the user's position changes.